--- client/src/routes/about/index.lazy.tsx
@@ -1,10 +1,0 @@
-import { createLazyFileRoute } from '@tanstack/react-router';
-
-export const Route = createLazyFileRoute('/about/')({
-  component: About,
-})
-
-function About() {
-
-  return <div className="p-2">Hello</div>
-}
