+++ client/src/routes/test/index.lazy.tsx
@@ -1,0 +1,5 @@
+import { createLazyFileRoute } from '@tanstack/react-router'
+
+export const Route = createLazyFileRoute('/test/')({
+  component: () => <div>Hello /test/!</div>
+})
