+++ vite.config.ts
@@ -13,7 +13,6 @@ export default defineConfig({
     vono(),
     tsconfigPaths(),
     TanStackRouterVite({
-      routeFileIgnorePrefix: "~",
       routesDirectory: "./client/src/routes/",
       generatedRouteTree: "./client/src/routeTree.gen.ts",
     }),
